stravamcpserver/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "strava-mcp"
version = "0.1.0"
description = "MCP server for Strava fitness data"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "stravalib>=1.0.0",
    "python-dotenv>=1.0.0",
]
